assignment 3/requirements.txt
# Core ML / Training
torch>=2.2.0
transformers>=4.43.0
peft>=0.11.0
trl>=0.9.0
bitsandbytes>=0.43.0
accelerate>=0.30.0
datasets>=2.19.0
einops>=0.8.0
flash-attn>=2.5.0; platform_system == "Linux"

# Evaluation metrics
rouge-score>=0.1.3
bert-score>=0.3.13
evaluate>=0.4.0

# API / data utilities
openai>=1.30.0
numpy>=1.26.0
pandas>=2.2.0
tqdm>=4.66.0
pyyaml>=6.0.1
jsonschema>=4.22.0
scipy>=1.13.0
scikit-learn>=1.5.0
